Adding changes.

+++ lib/src/main/java/data/Budget.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "budget_table")
 data class Budget(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo(name = "budget_amount") val budget: Float
+    @ColumnInfo(name = "budget_amount") val budget: Int
 )
