+++ lib/src/main/java/util/CalculateAmount.kt
@@ -1,6 +1,6 @@
 package util
 
-fun calculateAmount(operator: String, amount: Float): Float {
+fun calculateAmount(operator: String, amount: Int): Int {
     return if (operator == "+") {
         amount
     } else {
